migrate logic to wishlist talons with compatible extension
