+++ packages/peregrine/lib/talons/Wishlist/WishlistButton/useWishlistButton.ee.js
@@ -1,0 +1,8 @@
+import { useEffect } from 'react';
+
+export const useWishlistButton = props => {
+    useEffect(() => {
+        console.log(`I'm EE!`);
+    }, []);
+    return {};
+};
